requirements/base.txt
Django==1.8
PyJWT==1.1.0
django-cors-headers==1.0.0
django-filter==0.9.2
django-grappelli==2.6.4
django-model-utils==2.2
djangorestframework==3.1.1
djangorestframework-jwt==1.5.0
factory-boy==2.5.2
psycopg2==2.6.1
